#patch add reward

+++ middleware.ts
@@ -4,6 +4,7 @@ import { ACCESS_TOKEN_KEY } from "@/constants/token.constant";
 
 import dayjs from "@/lib/dayjs";
 import { deleteCookie } from "cookies-next";
+import { protectedRoutes } from "@/constants/route.constant";
 
 export function middleware(request: NextRequest) {
   const currentUser = request.cookies.get(ACCESS_TOKEN_KEY)?.value;
@@ -14,12 +15,17 @@ export function middleware(request: NextRequest) {
     pathname.startsWith("/api") // exclude API routes
   )
     return NextResponse.next();
-  if (currentUser) {
-    const { expiration_date } = JSON.parse(currentUser);
-    if (dayjs().isSameOrAfter(expiration_date)) {
-      deleteCookie(ACCESS_TOKEN_KEY);
+
+  if (protectedRoutes.includes(pathname)) {
+    if (currentUser) {
+      const { expiration_date } = JSON.parse(currentUser);
+      if (dayjs().isSameOrAfter(expiration_date)) {
+        deleteCookie(ACCESS_TOKEN_KEY);
+        return NextResponse.redirect(new URL("/", request.url));
+      }
+    } else {
       return NextResponse.redirect(new URL("/", request.url));
     }
-    return NextResponse.next();
   }
+  return NextResponse.next();
 }
